+++ launch/carla_service.launch.py
@@ -20,35 +20,27 @@
 #   OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 #   SOFTWARE.
 
-import os
-
-from ament_index_python.packages import get_package_share_directory
 import launch
+import launch_ros.actions
 
 
 def generate_launch_description():
     ld = launch.LaunchDescription([
-        launch.actions.DeclareLaunchArgument(
-            name='objects_definition_file',
-            default_value=get_package_share_directory(
-                'carla_autoware_bridge') + '/config/objects.json'
-        ),
-        launch.actions.IncludeLaunchDescription(
-            launch.launch_description_sources.PythonLaunchDescriptionSource(
-                os.path.join(get_package_share_directory(
-                    'carla_spawn_objects'), 'carla_example_ego_vehicle.launch.py')
-            ),
-            launch_arguments={
-                'objects_definition_file': launch.substitutions.LaunchConfiguration(
-                    'objects_definition_file'),
-            }.items()
-        ),
-        launch.actions.IncludeLaunchDescription(
-            launch.launch_description_sources.PythonLaunchDescriptionSource(
-                os.path.join(get_package_share_directory(
-                    'carla_autoware_bridge'), 'carla_service.launch.py')
-            ),
-        ),
+        launch_ros.actions.Node(
+            package='carla_autoware_bridge',
+            executable='carla_service',
+            name='carla_service',
+            on_exit=launch.actions.Shutdown(),
+            parameters=[
+                {
+                    'use_sim_time': True
+                },
+            ],
+            remappings=[
+                ('~/input/vehicle_status', '/carla/ego_vehicle/vehicle_status'),
+                ('~/output/control', '/carla/ego_vehicle/vehicle_control_cmd'),
+            ],
+        )
     ])
     return ld