+++ test/test_converters.py
@@ -23,18 +23,22 @@
 
 import struct
 
-from autoware_auto_control_msgs.msg import AckermannControlCommand
 from autoware_auto_vehicle_msgs.msg import SteeringReport, VelocityReport
+from carla_autoware_bridge.converter.actuation_status import ActuationStatusConverter
 from carla_autoware_bridge.converter.control_command import ControlCommandConverter
 from carla_autoware_bridge.converter.fake import FakeConverter
 from carla_autoware_bridge.converter.lidar_ex import LidarExtendedConverter
 from carla_autoware_bridge.converter.steering_status import SteeringStatusConverter
 from carla_autoware_bridge.converter.velocity_report import VelocityReportConverter
-from carla_msgs.msg import CarlaEgoVehicleControl, CarlaEgoVehicleStatus
+from carla_msgs.msg import (
+    CarlaEgoVehicleControl,
+    CarlaEgoVehicleStatus,
+    CarlaEgoVehicleSteering)
 from nav_msgs.msg import Odometry
 import numpy as np
 import pytest
 from sensor_msgs.msg import PointCloud2, PointField
+from tier4_vehicle_msgs.msg import ActuationCommandStamped, ActuationStatusStamped
 
 
 def test_fake_convert_without_setting_inbox():
@@ -107,19 +111,19 @@ def test_velocity_report_invalid_input():
 
 
 def test_left_steering_status_convert():
-    input_vehicle_status = CarlaEgoVehicleStatus()
-    input_vehicle_status.control.steer = -1.0
-
     fl_max_left_angle = -48.99
     fr_max_left_angle = -35.077
-    average_max_left_angle = (fl_max_left_angle + fr_max_left_angle) / 2
+    average_max_left_angle = np.radians(fl_max_left_angle + fr_max_left_angle) / 2
 
-    expected_steering_tire_angle = -np.radians(average_max_left_angle)
+    input_vehicle_steering = CarlaEgoVehicleSteering()
+    input_vehicle_steering.steering_tire_angle = average_max_left_angle
+
+    expected_steering_tire_angle = -average_max_left_angle
     expected_steering_status = SteeringReport()
     expected_steering_status.steering_tire_angle = expected_steering_tire_angle
 
     steering_status_converter = SteeringStatusConverter()
-    steering_status_converter.inbox = input_vehicle_status
+    steering_status_converter.inbox = input_vehicle_steering
     steering_status_converter.convert()
 
     assert pytest.approx(steering_status_converter.outbox.steering_tire_angle) == \
@@ -127,19 +131,19 @@ def test_left_steering_status_convert():
 
 
 def test_right_steering_status_convert():
-    input_vehicle_status = CarlaEgoVehicleStatus()
-    input_vehicle_status.control.steer = 1.0
-
     fl_max_right_angle = 35.077
     fr_max_right_angle = 48.99
-    average_max_right_angle = (fl_max_right_angle + fr_max_right_angle) / 2
+    average_max_right_angle = np.radians(fl_max_right_angle + fr_max_right_angle) / 2
 
-    expected_steering_tire_angle = -np.radians(average_max_right_angle)
+    input_vehicle_steering = CarlaEgoVehicleSteering()
+    input_vehicle_steering.steering_tire_angle = average_max_right_angle
+
+    expected_steering_tire_angle = -average_max_right_angle
     expected_steering_status = SteeringReport()
     expected_steering_status.steering_tire_angle = expected_steering_tire_angle
 
     steering_status_converter = SteeringStatusConverter()
-    steering_status_converter.inbox = input_vehicle_status
+    steering_status_converter.inbox = input_vehicle_steering
     steering_status_converter.convert()
 
     assert pytest.approx(steering_status_converter.outbox.steering_tire_angle) == \
@@ -147,14 +151,14 @@ def test_right_steering_status_convert():
 
 
 def test_center_steering_status_convert():
-    input_vehicle_status = CarlaEgoVehicleStatus()
-    input_vehicle_status.control.steer = 0.0
+    input_vehicle_steering = CarlaEgoVehicleSteering()
+    input_vehicle_steering.steering_tire_angle = 0.0
 
     expected_steering_status = SteeringReport()
     expected_steering_status.steering_tire_angle = 0.0
 
     steering_status_converter = SteeringStatusConverter()
-    steering_status_converter.inbox = input_vehicle_status
+    steering_status_converter.inbox = input_vehicle_steering
     steering_status_converter.convert()
 
     assert pytest.approx(steering_status_converter.outbox.steering_tire_angle) == \
@@ -171,15 +175,20 @@ def test_steering_status_invalid_input():
         steering_status_converter.convert()
 
 
-def test_throttle_control_command():
-    input_control_command = AckermannControlCommand()
-    input_control_command.longitudinal.acceleration = 1.842
+@pytest.fixture
+def control_command_converter():
+    steer_map_path = './test/data/steer_map.csv'
+    return ControlCommandConverter(steer_map_path)
+
+
+def test_throttle_control_command(control_command_converter):
+    input_control_command = ActuationCommandStamped()
+    input_control_command.actuation.accel_cmd = 0.301
 
     expected_control_command = CarlaEgoVehicleControl()
-    expected_control_command.throttle = 1.0
+    expected_control_command.throttle = 0.301
     expected_control_command.brake = 0.0
 
-    control_command_converter = ControlCommandConverter()
     control_command_converter.inbox = input_control_command
     control_command_converter.convert()
 
@@ -189,15 +198,14 @@ def test_throttle_control_command():
         pytest.approx(expected_control_command.brake)
 
 
-def test_still_control_command():
-    input_control_command = AckermannControlCommand()
-    input_control_command.longitudinal.acceleration = 0.0
+def test_still_control_command(control_command_converter):
+    input_control_command = ActuationCommandStamped()
+    input_control_command.actuation.accel_cmd = 0.0
 
     expected_control_command = CarlaEgoVehicleControl()
     expected_control_command.throttle = 0.0
     expected_control_command.brake = 0.0
 
-    control_command_converter = ControlCommandConverter()
     control_command_converter.inbox = input_control_command
     control_command_converter.convert()
 
@@ -207,15 +215,14 @@ def test_still_control_command():
         pytest.approx(expected_control_command.brake)
 
 
-def test_brake_control_command():
-    input_control_command = AckermannControlCommand()
-    input_control_command.longitudinal.acceleration = -6.168
+def test_brake_control_command(control_command_converter):
+    input_control_command = ActuationCommandStamped()
+    input_control_command.actuation.brake_cmd = 0.205
 
     expected_control_command = CarlaEgoVehicleControl()
     expected_control_command.throttle = 0.0
-    expected_control_command.brake = 1.0
+    expected_control_command.brake = 0.205
 
-    control_command_converter = ControlCommandConverter()
     control_command_converter.inbox = input_control_command
     control_command_converter.convert()
 
@@ -225,18 +232,13 @@ def test_brake_control_command():
         pytest.approx(expected_control_command.brake)
 
 
-def test_steering_left_control_command():
-    fl_max_left_angle = 48.99
-    fr_max_left_angle = 35.077
-    average_max_left_angle = (fl_max_left_angle + fr_max_left_angle) / 2
-    expected_steering_tire_angle = np.radians(average_max_left_angle)
-    input_control_command = AckermannControlCommand()
-    input_control_command.lateral.steering_tire_angle = expected_steering_tire_angle
+def test_steering_left_control_command(control_command_converter):
+    input_control_command = ActuationCommandStamped()
+    input_control_command.actuation.steer_cmd = -0.648202002048492  # rad
 
     expected_control_command = CarlaEgoVehicleControl()
-    expected_control_command.steer = -1.0
+    expected_control_command.steer = 0.6
 
-    control_command_converter = ControlCommandConverter()
     control_command_converter.inbox = input_control_command
     control_command_converter.convert()
 
@@ -244,18 +246,13 @@ def test_steering_left_control_command():
         pytest.approx(expected_control_command.steer)
 
 
-def test_steering_right_control_command():
-    fl_max_right_angle = -48.99
-    fr_max_right_angle = -35.077
-    average_max_right_angle = (fl_max_right_angle + fr_max_right_angle) / 2
-    expected_steering_tire_angle = np.radians(average_max_right_angle)
-    input_control_command = AckermannControlCommand()
-    input_control_command.lateral.steering_tire_angle = expected_steering_tire_angle
+def test_steering_right_control_command(control_command_converter):
+    input_control_command = ActuationCommandStamped()
+    input_control_command.actuation.steer_cmd = 0.340341448783875  # rad
 
     expected_control_command = CarlaEgoVehicleControl()
-    expected_control_command.steer = 1.0
+    expected_control_command.steer = -0.3
 
-    control_command_converter = ControlCommandConverter()
     control_command_converter.inbox = input_control_command
     control_command_converter.convert()
 
@@ -263,13 +260,12 @@ def test_steering_right_control_command():
         pytest.approx(expected_control_command.steer)
 
 
-def test_gear_control_command():
-    input_control_command = AckermannControlCommand()
+def test_gear_control_command(control_command_converter):
+    input_control_command = ActuationCommandStamped()
 
     expected_control_command = CarlaEgoVehicleControl()
     expected_control_command.manual_gear_shift = False
 
-    control_command_converter = ControlCommandConverter()
     control_command_converter.inbox = input_control_command
     control_command_converter.convert()
 
@@ -277,11 +273,10 @@ def test_gear_control_command():
         pytest.approx(expected_control_command.manual_gear_shift)
 
 
-def test_control_comman_invalid_input():
+def test_control_comman_invalid_input(control_command_converter):
     class UnexpectedInput():
         pass
     input_invalid = UnexpectedInput()
-    control_command_converter = ControlCommandConverter()
     control_command_converter.inbox = input_invalid
     with pytest.raises(RuntimeError):
         control_command_converter.convert()
@@ -486,3 +481,36 @@ def test_lidar_ex_invalid_input():
     lidar_ex_converter.inbox = input_invalid
     with pytest.raises(RuntimeError):
         lidar_ex_converter.convert()
+
+
+def test_actuation_status():
+    input_vehicle_status = CarlaEgoVehicleStatus()
+    input_vehicle_status.control.throttle = 0.4
+    input_vehicle_status.control.brake = 0.0
+    input_vehicle_status.control.steer = -0.1
+
+    expected_actuation_status = ActuationStatusStamped()
+    expected_actuation_status.status.accel_status = 0.4
+    expected_actuation_status.status.brake_status = 0.0
+    expected_actuation_status.status.steer_status = -0.1
+
+    actuation_status_converter = ActuationStatusConverter()
+    actuation_status_converter.inbox = input_vehicle_status
+    actuation_status_converter.convert()
+
+    assert pytest.approx(actuation_status_converter.outbox.status.accel_status) == \
+        pytest.approx(expected_actuation_status.status.accel_status)
+    assert pytest.approx(actuation_status_converter.outbox.status.brake_status) == \
+        pytest.approx(expected_actuation_status.status.brake_status)
+    assert pytest.approx(actuation_status_converter.outbox.status.steer_status) == \
+        pytest.approx(expected_actuation_status.status.steer_status)
+
+
+def test_actuation_status_invalid_input():
+    class UnexpectedInput():
+        pass
+    input_invalid = UnexpectedInput()
+    actuation_status_converter = ActuationStatusConverter()
+    actuation_status_converter.inbox = input_invalid
+    with pytest.raises(RuntimeError):
+        actuation_status_converter.convert()